CORS Exception middleware and Extension method (#7)
* DI extension method and CORS configuration

* Add ExceptionHandler middleware

* Void methods

* Cleaning

+++ Pets.BLL/Services/PetService.cs
@@ -3,11 +3,6 @@ using Pets.BLL.Interfaces;
 using Pets.BLL.Models;
 using Pets.DAL.Entities;
 using Pets.DAL.Interfaces;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Pets.BLL.Services
 {
@@ -30,7 +25,12 @@ namespace Pets.BLL.Services
 
         public Pet GetById(int id)
         {
-            return _mapper.Map<PetEntity, Pet>(_repository.GetById(id));
+            PetEntity pet = _repository.GetById(id);
+
+            if(pet == null)
+                throw new ArgumentNullException("Object doesn't exist");
+            else
+                return _mapper.Map<PetEntity, Pet>(pet);
         }
 
         public IEnumerable<Pet> Find(Func<Pet, bool> predicate)
@@ -51,7 +51,12 @@ namespace Pets.BLL.Services
 
         public void Delete(int id)
         {
-            _repository.Delete(id);
+            PetEntity pet = _repository.GetById(id);
+
+            if(pet == null)
+                throw new ArgumentNullException("Object doesn't exist");
+            else
+                _repository.Delete(id);
         }
     }   
 }
